Windows installer: Update SBCL (only for 64 bit).
The SBCL team did not release a 32 bit installer.

Building the installer works with the included wine 5.0
in Ubuntu 20.04, the testsuite fails with a stack overflow.

But it does work on Windows 10 and it does work, when
using wine staging 7.16 form winehq.org, so I updated
SBCL.
Maybe I should also update my development machines soon.