Exit building of docs if makeinfo fails
Basically, makeinfo should not have errors building the docs and if
there are, we should exit the build so errors aren't buried in a ton
of output.

First, update rule in Makefile.am for building maxima.info.  Remove
"-" so that makeinfo errors are not ignored.

Next, in build_html.sh.in, remove the makeinfo options "--force -e
10000" because makeinfo should succeed these days.  No need to skip
over the errors.  It's best if we keep makeinfo working without
errors.

Finally, as a simple change bracket the makeinfo commands in
build_html.sh.in with set -x/set +x so we can see the commands that
are being run.