HTML help: Use the default browser on Windows.
start <url>
(without the name of a browser)
calls the default browser with the URL <url>.

Mention, how one can set a non default browser on
other operating systems.

The case of the format string (~a or ~A) is not
relevant, correct? Use the same case everywhere, not
... format string of the form @code{<cmd> ~a} ...
and later
... the default setting is @code{"start msedge -url ~A"}.