Add some more info for hompack
We forgot to describe what some of the optional keyword args do, so
update the docs for that.   Also some clean of the description of the
return values to mention that these are lists.

Finally, in the first example, load hompack, so the user knows how to
get hompack_polsys.