Add missing anchors
First, modify build_html.sh so that we don't force makeinfo (comment
out "FORCE=--force" for ru).

Run make and look at the nonexistent cross references.  This change
adds all the missing anchors so that the refs exist.  However, there
are still many missing anchors because the Russian documents don't
include the missing items.  This is caused by the documents
incorporating the English versions of the the texi files (like
orthopoly.texi, simplex.texi, etc.) which are newer than when the
Russian translation was written.

Sadly it even seems as if "log" is not a defined function.