Add xref to "Getting Started with ODEPACK"
In the description for dlsode_step and dlsode_init, add a cross-ref to
the section "Getting Started with ODEPACK" to help users figure out
where to go to learn how to use the functions.

Also fixed a typo, and added @math{} to a couple of places to make it
typeset a bit nicer.