Recognize and use user-specified init files
If the user specifed a mac or lisp init file (via --init-mac or
--init-lisp), use the given file for initialization.  This is searched
for as normal for $load and $batchload.  An error is signaled if the
file can't be found, but we just continue since there's not much we
can do during init about it.

If no user-specified init file is given, load the default file in the
maxima userdir, if found, as we were before.

This is implemented by setting *maxima-initmac* and *maxima-initlisp*
to NIL and is set from the corresponding command-line option.  The
default values are now in *defaut-maxima-initmac*
and *default-maxima-initlisp*.