Simplify handling of --init and --init-mac,--init-lisp
The arg of init now has any directory parts removed, and the last
file name extension is also removed.  The resulting basename is to
form the .mac and .lisp init files that are now only searched for in
the userdir.

This is a breaking change in how --init works;  Previously no
stripping of the directory or extension was done.

Also --init-mac and --init-lisp are now aliases for --preload and
behave exactly like --preload.  This is a small change from before
where multiple --init-mac options would only use the last value.  Now
each of these will load the specified file.

Clean up macys.lisp a bit since there isn't a difference between
*maxima-initmac* and *maxima-initlisp* no longer hold anything.