MEVALP_TR: return result of MEVALP1_TR instead of unknown
If ERROR? is NIL, return the result from MEVALP1_TR instead of unknown.
It will be useful to have MEVALP_TR return the result in later commits.

This also makes MEVALP_TR behave more like MEVALP.

The behavior of IS-BOOLE-CHECK and MAYBE-BOOLE-CHECK have not changed.
They now call a different common function to get the desired symbol
unknown when the result is not a boolean.

There is currently only one use of MEVALP_TR outside of the translator
(in nset).  This change does not affect that because ERROR? is T in
that case, so unknown would never be returned.

See also commit baf41788.

No problems with the test suite, share test suite or rtest_translator.