Consolidate code for evaluating and/or/not expressions
The functions IS-MAND_TR, IS-MOR_TR and IS-MNOT_TR used by translated
code are almost identical to the IS-MAND, IS-MOR and IS-MNOT used by
the interpreter.  The only significant difference is what functions
they call to evaluate predicates.

Now IS-MAND and friends take a parameter specifying how to evaluate
predicates.  The interpreter can pass MEVALP and the transl code can
pass MEVALP_TR, so IS-MAND_TR and friends can be removed.

No problems with the test suite, share test suite or rtest_translator.