Fix the bogus translations of <= and >=
The translation of <= has been similar to > followed by not (TRP-NOT),
and similarly the translation of >= has been similar to < followed by
not.  This is wrong:

(%i1) foo () := is (%i <= 1)$

(%i2) foo ();
(%o2) false

(%i3) translate (foo)$

(%i4) foo ();
(%o4) true

Also the translator basically assumed that prederror would be true.
The way the not used here was implemented (TRP-NOT), it would behave
like NOT when given a boolean value but otherwise it would simply
return its argument unchanged.  This is not a problem for is/maybe,
but after an upcoming commit (to make if honor prederror) this could
lead to bogus results in conditionals when prederror is false.

Now we translate <= and >= directly, using TRP-INEQUALITY (introduced
in commit 7f12bb7f).

No problems with the test suite, share test suite or rtest_translator.
New tests have been added to rtest_translator.