Improve the translation of notequal
The translation of notequal has been similar to equal followed by
not (TRP-NOT).  Now we translate notequal more directly, using MNQP
in the general case.

The equal followed by not is not a problem for is/maybe, but the way
TRP-NOT is implemented would be a problem in the future.  An upcoming
commit will make if honor prederror, and TRP-NOT could yield bogus
results when prederror is false.

See also commit cf7c667c for more.

No problems with the test suite, share test suite or rtest_translator.
New tests have been added to rtest_translator.