Change the translation of # to avoid the problematic TRP-NOT
The translation of # is the last user of TRP-NOT (see commits cf7c667c
and f7b46967 for more on TRP-NOT).

Instead of wrapping the translation of = in a call to TRP-NOT, let's
just translate # as the logical negation of =.

Then we can just remove TRP-NOT.

No problems with the test suite, share test suite or rtest_translator.
New tests have been added to rtest_translator.