In Gnuplot output, format title string with ~a (and enclose in double quotes) instead of ~s.
With ~a, backslash escape sequences such as \n are printed as such (with one backslash),
but with ~s, two backslashes are printed (presumably to make the result readable as
an identical string). Example:

* (format t "\"~a\" ---- ~s~%" "foo\\nbar" "foo\\nbar")
"foo\nbar" ---- "foo\\nbar"

In order for Gnuplot to correctly interpret the \n, there must be just one backslash.