Create globals.lisp to hold global vars for Maxima
First cut at adding globals.lisp to hold global
variables (defvars/defmvars) that are used in multiple files in
Maxima.  This is done to make it clearer that these are used in more
than one file and to simplify specifying the dependencies the build
system.

We move the defvars `infinities`, `real-infinities`, and
`infinitesimals` to globals.lisp.

Files are updated accordingly.

maxima.system and maxima.asd are updated by explicitly mentioning the
dependency on globals.lisp in the modules that contained the files we
needed to modify.