Have defmvar recognize some options
Historically some defmvar definitions included options as demonstrated
by
```
(defmvar $linenum 1 "the line number of the last expression."
	 fixnum no-reset)
```
The options `fixnum` and `no-reset` (and any others) were ignored.

Modify `defmvar` to recognize these.  `no-reset` causes the expansion
not to add the variable to `*variable-initial-values*` so that the
variable will not be reset when `reset()` is called.

If `fixnum` or `boolean` is given declaim the variable as having that
type.

This allows us to change `$activecontexts` to be a `defmvar` with a
`no-reset` option.  The testsuite passes now.  Before, changing
`$activecontexts` to a `defmvar` causes the test suite to fail,
presumably because it was reset at various points in the test.