Defmvar supports setting properties
Update defmvar to support the option `:properties` which has the from
`((ind1 val1) (ind2 val2) ...)`, which sets the property `ind1` to the
value `val1` and so on for the variable.

Updated a couple of places in this file where there was a separate
putprop to define the property in the defmvar.