Process defmvar options more carefully
We want to process the options list more carefully by looking at each
option and handling them appropriately.  If an option is not
recognized, we signal an error.

This allowed us to find all the options that are currently used and
explicitly ignore them if we don't want to handle them.  Thus, we have
a record of all the options that are currently used.  At some later
date we can figure out what to do with them, if it proves helpful.

Also, we support `properties` as an alias of `:properties`.  We prefer
`:properties`.