Set evflag property via defmvar in globals.lisp
In mlisp, the evflag property is set for many symbols defined in
globals.lisp.  Instead, set this property via defmvar in globals.lisp,
and remove the variable from the list in mlisp.lisp.

There are still a few evflag values, but these variables are defined
elsewhere which we won't convert now.