This commit fixes the bugs
 #3592 Wrong limit
 #3415 limit doesn't check for zero coefficients in limit((a*x+1)/(a*x+2),x,inf)
 #3279 limit incorrect with domain:complex
 #3136 gruntz(atan2(x^2 - 2, x^3 - 2*x), x, sqrt(2), minus) => atan2(0,0) undefined
 #3280 gruntz incorrect limit

Additionally, it allows Maxima to find the limit of fib expressions. The new function extra-simp dispatches
radcan on expressions of the form (positive integer)^X and it replaces every atom that has been declared
to vanish by zero. The other fix involves the L'Hopital rule--it was testing for a vanishing expression by
equality to zero--this patch changes it to test using mnqp. This fixes bug 3592.