Fix testsuite failures with sbcl due to defmvar decls
The defmvar macro recognizes type declarations and adds a type
declaration when found.  However, in the testsuite there are many
failures because the the value is not of the specified type.

For now, we'll just recognize the declarations, but ignore them.  At
some later point, we'll need to investigate these failures and decide
if we really want to have the declarations.

Thanks to Barton for pointing this out to me.