Support defmvar option :setting-predicate
Support `:setting-predicate` (and `setting-predicate`) option in `defmvar`.
We prefer `:setting-predicate`, however.

Update `$gensumnum` and `$genindex` to support this.

The predicate for `$gensumnum` was not correct, so it's been updated to
verify the value must either be NIL or a non-negative integer.  This
also required changing `gensumindex` to allow `$gensumnum` to be NIL.