Move set of evfun property to the function
Modify `defmfun` to allow either a symbol or a list defining the
function.  The list looks like (name :properties p) where `name` is the
name of the function and `p` is a list of lists for the properties that
we wish to assign to `name`.  Similar to the `:properties` option for
`defmvar`.

Remove the `mapc` that set the 'evfun property on a list of functions,
and use the new :property feature of defmfun to set the property when
the function is defined.