Move remaining sets of evfun property to the function.
The remaining cases of 'evok were for defmspec's.  Thus, modify
defmspec to allow setting properties in the definition.  This is
basically the same syntax as for defmfun.

Use the new defmspec to define the evok property with the definition
and remove the mapc call in mlisp.lisp that set these.