Move 'assign property for $ratvars from msetchk to defmvar.
First, change defmvar :properties so that user's have to quote any
symbols if needed.  This allows the 'assign property to be generated
with a lambda expression.

Second update all uses of `:properties ((assign foo))` to
`:properties ((assign 'foo))`.

Finally, for $ratvars, use a lambda expression for the 'assign
property and remove the corresponding code from `msetchk`.

Testsuite passes.