Move putprop assign to defmvar :properties
In globals.lisp, there were a few calls to `putprop` to set the
`assign property.  Move these to the `:properties` option of the
corresponding `defmvar`.

Also for `posintegerset`, make the error message clearer because the
comments say "a positive integer, including zero", but the message
says "a positive integer".  Replace that with "non-negative integer".