Rewrite to remove ''MAKE-FUN and ''MAKE
Get rid of the function ''MAKE-FUN and just move it to the macrolet
''MAKE.  And rename ''MAKE to just `make-fun` since it's in a
macrolet, we don't need weird names.

And since we use the macro `ftake` from mopers.lisp, update dependency
to make sure mopers is loaded before mlisp.