Add defmvars for variables in $infolists
Many of the variables listed in `$infolists` aren't defined anywhere,
but they're clearly special variables.  Thus define them here and set
the appropriate options for them.

One tweak is that `$structures` was a defvar, which we've replaced
with `defmvar` with the `no-reset` option.

Then we can remove the mapc in nisimp that initalized all of the
variables to `(mlist simp)`.

This allows us to remove the mapc that sets the 'assign property to
neverset in suprv1.lisp.

And finally, we fix an inconsistency because suprv1.lisp set all the
variables in `$infolists` to have an 'assign property of `'neverset`.
But nisimp.lisp sets the propert to `'let-rule-setter`.  This latter
setting appears to be the actual value in use in maxima.