Remove the commented out mapc's
Remove the commented out mapc's in mlisp.lisp and suprv1.lisp.  The
initialization done in mlisp.lisp is now in the defmvars.

Setting the 'assign property is no longer needed in suprv1.lisp
because it's also done in the defmvars.  (Besides, it incorrectly set
let_rule_packages to 'neverset, which appears to be wrong.)