Fix up some compiler warnings about undefined vars.
Ecl warns about undefined variables in init-cl and macys.  Create
defmvars for these in globals.lisp.

Also use the defmvar :properties feature to set some properties.  This
allows us to remove the `def-lisp-shadow` macro and calls in
init-cl.lisp.

Not sure why we need the property 'lisp-shadow when it seems that
'alias would do the trick.  Perhaps that's for another time.  It would
be nice to have one less property.