Fix compiler note about undefined Z in def-simp expintegral_e
When moving the simplifier for `expintegral_e` to use
`def-simplifier`, we lost the `z` arg, which is now some unsused
gensym.  But it was used in this simplifier as a temp.  So add a local
var `z` at the appropriate places so it works and we don't get a
compiler warning.