Texi init file for texinfo 7 and later
This is the init file to cause makeinfo to produce html files of the
form maxima_<nnn>.html, as we've always done.  The original version
used undocumented functions to do that.  This version now uses the
public API to do this.

This was written by Patrice Dumas and posted to the help-texinfo
mailing list on 2022-10-19.