Move modulus aliases from suprv1 to globals.
Instead of defining the aliases betweeen `modulus` and `$modulus` in
suprv1.lisp, add it to globals.lisp, right after `modulus` is defined
so that it's easy to see that these are aliases of each other.

Add a note that we should probably replace all uses of `modulus` with
`$modulus` and get rid of the aliases.