Minor tweak to make it easier to read build_info.
Instead of having the reader conditional in `build_info` for sbcl and
not sbcl, always call `ensure-readably-printable-string` for all
lisps.  Put the impl-specific details in
`ensure-readably-printable-string`.  For other lisps, this function
just returns the arg.  Since getting the build info isn't a
speed-critical item, an extra function call for other lisps shouldn't
matter.