Refactor dimension-build-info to reduce common expressions.
There was a lot of duplicated code in `dimension-build-info`
for creating the string to be printed and formatting it for display.
The only difference was the label to use and the symbol to be printed.

Put all of this into one function that does both to eliminate the
redundant code.