Get rid of warnings about undefined vars old-ibase and old-base.
Just add defvars for these, but also rename them to `*old-ibase*` and
`*old-base*` to follow CL conventions for special variables to make it
clear they're special.

It would probably be better to bind `*print-base*` and `*read-base*`
to 10 as needed instead of doing it this way.  There are already a few
cases for `*print-base*`.