Remove translator vars from lmdcls.
These are either defined in transl before use or are only used in one
file.  Also update dependencies so that transl is compiled before the
dependent files.

Finally, add comment that transm.lisp references the special var
`tr-abort` in transl.lisp.  But we can't add that
`maxima-language-compiler-macros` depends on
`maxima-language-compiler`.