Remove translator vars from lmdcls.
These are all defined and then used in one file except for `tr-abort`
and `transl-file`.

`tr-abort` was handled in a previous commit, but we can remove it from
lmdcls, because transm.isp declares it special before using it.

`trans-file` is defined in transs, but also used in only transm (and
nowhere else), which has a special decl for it.  Thus, it can be
removed from lmdcls.