Remove sums from lmdcls
It's not really clear if `sums` (from simp.lisp) should really be
special; it's a common word and used in lots of places.  From what I
can tell, those uses aren't related to the variable `sums` from
simp.lisp.  This is further reinforced by the fact that in an early
commit, `sums` was renamed to `*sums*` in simp.lisp without affecting
the test suite.

Removing this from lmdcls didn't change the test suite results; the
test suite passed for both cmucl and ecl with the expected results.