Remove meta-prop-l and meta-prop-p from lmdcls
These two vars are defined in trpop.lisp and are only used there and
in trans4.lisp.  Since these are defined by defmvar, instead of moving
these to globals.lisp, let's just tell defsystem that trans4.lisp
depends on trprop.

(Perhaps it would make sense to move these to transl which more or
less contains most of the special vars for the translator.  For
another day.)