Remove $maperror and $maxapplyheight from lmdcls
`$maperror` is defined in mlisp.lisp and only used there and once in
trdata.lisp to set 'mode property to $boolean for the translator.

`$maxapplyheight` is defined in matrun.lisp and only used there and
once in trdata.lisp to set 'mode property to $boolean for the
translator.

Thus, just make sure mlisp.lisp is compiled before trdata.lisp in
maxima.system.

Finally, in matrun.lisp, since `$maxapplyheight` and `$maxapplydepth`
are defmvars, we don't need the declare-top special for them, so
remove that.