Fix compiler warning about undefined $opproperties
First, move the `defparameter` for from the end of the file and change
it from `defparameter` to `defmvar` with `no-reset`.  Also make it so
that the user can't assign values to it directly.

The initial value is obtained by looking at the values from a previous
version of maxima.  But it could also be obtained by looking at the
value of `opers` and reversing the elements.

Testsuite passes.