Add more info for psi function.
Add links to maxpsiposint, maxpsinegint, maxpsifracnum, and
maxpsifracdenom in the description of the function psi so the user of
the html doc can find them easily.

For each variable maxpsiposint, maxpsinegint, maxpsifracnum, and
maxpsifracdenom, use m4_math to make it a bit nicer to read in html,
and also add some examples on what impact these variables have on
computing psi[n](x).

For example, maxpsifracnum and maxpsifracdenom only affect psi[0](x).
But maxipsiposint and maxpsinegint affect all orders n.