Make equations be lexical.
The variable `equations` has a declare-top special in mat.lisp,
and polyrz.lisp.  We can make equations lexical by having `solve2` in
mat.lisp take a second parameter `equations` and having it return a
second value that is `equations`.

Then in polyrz.lisp, all calls to solve2 are updated to add a nil
second argument and to save the second return value to the lexical var
`equations`.

Note that the use of `equations` only happens when `programmode` is
false.  Some basic tests for this works.

Full test suite passes.