Clear up description of factorial_expand
Previously it said `factorial_expand` controlled simplification
of (n+1)! for integer n.  But that's not actually what happens and is
confusing if n is an integer, (n+1)! is alway simplified to a number.

Instead it controls the simplification (expansion) of (x+n)! when n is an
integer.  So (x+1)! is expanded to (x+1)*x!.

Clarify this.