Add TeX formulas for the halfangles description
Use TeX formulas for the half-angle formulas to make it nicer to read
in html.  We also made the text version a little simpler by writing
z=x+%i*y so that we can get rid of the realpart and imagpart
expressions in the text form of the complicated factor.