Move the halfangles section
The description of halfangles shows up between the description for
csch and ntrig (which is followed by sec).

That's a strange place for it.  Move the description after tan and
before trigexpand.  It seems a bit better to keep all the trig
functions together.

This really only matters if you're reading the html or pdf docs or the
info doc.  This isn't a problem for Maxima's text help since you only
see the section and not surrounding sections.