Move the atrig1 and ntrig sections
The description of atrig1 and ntrig show up between the description
for the trig functions.  That is rather odd.  Move the down after tan
and before halfangles.

This really only matters if you're reading the html or pdf docs or the
info doc.  This isn't a problem for Maxima's text help since you only
see the section and not surrounding sections.