Use def-simplifier for hypergeometric function
Modify def-simplifier to add :skip-simpcheck option so that the
generated code doesn't automatically call simpcheck on the args.  Also
change the Z arg (simp flag) not to be a gensym but a fixed name so
that the body can use the simp flag if needed.

Change simp-hypergeometric to use def-simplifier.

This isn't working yet.  rtest_hg fails.