More tweaks to get testsuite to pass
Need to define a `$hypergeometric` function so that when
hypergeometric is autoloaded, we replace the autof definition with a
real `$hypergeometric` is loaded so we don't autoload anymore.

So that we autoload when necessary, modify the places where we were
doing `(ftake '%hypergeometric ...)` to `($hypergeometric ...)` to
autoload hypergeometrics.  Without, we're left with the noun form
because that's not enough to autoload hypergeometrics.