In package descriptive, define some descriptive statistics with optional weight argument.
The functions which recognize the optional weight argument are:
mean, noncentral_moment, central_moment, var, std, var1, and std1.

The weight argument may be a literal 1 (not a list), indicating equal
weight for all data, or a list of length equal to the number of data.
Weights cannot be negative, and the sum of weights cannot be zero.
The weights, if a list, need not sum to 1; the weights are normalized
by their sum in internal calculations.

When weights are integers, the effect is exactly the same as duplicating
each datum the number of times given by the corresponding weight.

This commit includes unit tests for the modified functions.