Convert $safe_float to lisp and move to hypergeometric.lisp
`$safe_float` is only used in hypergeometric.lisp so convert to
lisp (which is relatively simple), and add it to hypergeometric.lisp.
Then remove the definition from hypergeometric_mac.mac.

This also means to remove the autoload for `$safe_float`.

Note:  The new definition of `$safe_float` is slightly different.
Previously, if you did `safe_float(2^2000)`, you would get a message
about 2^2000 not fitting in a double-float and you'd get a bfloat
result.  Now you silently get a bfloat result.