Fix #4077: replace opcons with ftake*
As mentioned in bug #4077, we replace calls of `opcons` in nummod.lisp
with the equivalent `ftake*`.  And also change `opcons` to use
`ftake*` to make the equivalence clearer.

The share directory has files using `opcons`, so we can't remove
`opcons` without fixing those.  I'm lazy and leaving them as is,
although it would be better if we didn't have two ways to do the same
thing.