Remove *in-compile* from lmdcls.
It's used only in transl.lisp, so might as well a `defvar` for it and
remove it from lmdcls.lisp.

Also add a comment that it appears that nothing sets `*in-compile*` to
anything, so maybe we can remove the var completely and the only other
use of it (in a `cond` case)?

Full testsuite passes.