Rename local to *local*
`local` is a defvar that appears to be used only in transl.lisp (Hard
to tell from a grep; it's a common word.)  So rename it to `*local*`
and remove it from lmdcls.lisp.

Full testsuite passes.  But I guess there aren't really any tests for
translate in the testsuite, so it's hard to know for sure.  At least
this change doesn't affect other parts of Maxima.