Rename alpha to *alpha*
`alpha` is used in a lot of places, and from what I can tell, many of
those places (like bessel.lisp and ellipt.lisp) don't intend to use
the `alpha` specvar.  (It was originally defined in ifactor.lisp).
So, let's rename it to `*alpha*` per lisp conventions and remove it
from lmdcls.lisp.  Then update code that actually wants to use the
specvar `alpha` to use `*alpha*`.

I think I got all the places that need to be changed.

The full testsuite passes.