Move *bigprimes* and *alpha to globals.lisp (from rat3c)
`*alpha` is a specvar used in several files so it makes sense to have
it in globals.lisp.  Since `*alpha` is initialized from `*bigprimes*`,
it makes sense to have this specvar in globals.lisp too.

However, `*bigprimes*` is initialized using the `next-prime` function
which isn't available when globals.lisp is compiled.  Hence, it has to
be initialized at startup in `initialize-runtime-globals`.

`*alpha` is initialized to the first value in `*bigprimes*` so we need
to initialize it in `initialize-runtime-globals` too.  Since `*alpha`
is a `defmvar`, we need to update the initial values hashtable too.

Finally, `$pointbound` is initialized to `*alpha`, so initialize it in
`initialize-runtime-globals` too.  The initial values hashtable needs
to be updated too.

Full testsuite passes.