Rename "Mathematical Functions" to "Elementary Functions"
As discussed earlier on the mailing list, rename "Mathematical
Functions" section to "Elementary Functions" since that is closer to
what this section is about and also serves to distinguish it from the
special functions and elliptic functions section, which are also
mathematical functions.

Also add some Up pointers to the sections.  Previously, we couldn't
navigate up to the "Elementary Functions" sections.  That was annoying
for navigation in the HTML doc.