Change Documentation Categories section to be an appendix
It was strange to see "Appendix A Function and Variable Index"
followed by "B Documentation Categories".  Hence, change it from a
`@chapter` to `@appendix` to make it "Appendix B Documentation
Categories".

Also update `@node links to so that the previous node of
"Documentation Categories" is the function and variable index.