Rename ans to *schatc-ans* in schatc.lisp
schatc.lisp has a `declare-top` special for `ans`, and a corresponding
unspecial the bottom.  This kind of makes it look like `ans` is only
special in this file.  Thus, rename `ans` to `*schatc-ans*`.

That, however, causes failures in the test suite, because it's
actually used in other places.  It's hard to know exactly where since
"ans" is such a common variable name.  This change renames `ans` in
series.lisp and trgred.lisp.

With this change, the full testsuite passes.