Replace specvar *p* with lexical var
The specvar `*p*` can be replaced by a lexical var by passing in the
value of `*p*` to functions that need it.  Then we can get rid of the
special declarations, replacing the specvar with a lexical var.

Rename `*p*` to `combin-par` to make it somewhat unique so as not to
clash inadvertently with some other possible specvar defined
elsewhere.