*i does not need to be special
`subsum` has the first arg of `*i` and it immediately calls `susum1`,
which is the only place to use `*i`.  We can make it lexical by adding
`*i` as an arg to `susum1`.  This requires a minor rewriting of the
`recur-apply` lambda.

Now we can remove the `declare-top` `special` for `*i`.

Full testsuite passes.