Change specvar *x* to a lexical var
`*x*` is a specvar that is used in a couple of places.  We can replace
the specvar by a lexical var by changing `pdecomp*` to take an extra
arg for the (former) specvar.  Use `combin-x` as the name to make it
less likely to collide with any other possible var.

Full testsuite passes.