Rename *var* to *combin-var*
`*var*` is used only in combin.lisp and hypgeo.lisp.  A comment in
hypgeo.lisp says that `*var*` is global only for that file.  Let's
make it so by renaming `*var*` in combin.lisp to `*combin-var*` to
make it definitely different from the one in hypgeo.lisp.  Then there
can't be any confusion on whether they're the same or not.

Full testsuite passes.