Rename *ans* to *combin-ans* in combin.lisp
Just to prevent any future possible confusion, rename `*ans*` to
`*combin-ans*` and make it a defvar.  We can remove the `declare-top`
special and unspecial for this.  (Note:  combin.lisp is the only place
I can find that uses `*ans*`, but let's make it unique since "ans" is
such a common variable name.)