Rename *a* to *combin-a*
It's not really clear if `*a*` is the same special var as `*a*` in
other files, but the unspecial at the end hints that it's intended to
be special only in this file.

Therefore, rename `*a*` to `*combin-a*` to make it clear that it's
not.  This change did not affect the full testsuite so I'm going to
assume it really is different.

Also, add a `defvar` for it and remove the unspecial declaration at
the end of the file.