Remove declare-top unspecial for w
Not sure why it's there.  Removing it has no effect on the full
testsuite, so it seems we don't need it.

A quick grep shows there are lots of "w"'s all over, but it doesn't
seem to be special, but it's hard to tell because there are so many.
`(find-symbol "W" "MAXIMA")` says there is such a variable.  Possibly
used as a pattern with the schatc matcher?