Rename specvar sum and usum
Let's rename the specvar `sum` and `usum` to `*combin-sum*` and
`*combin-usum*` to make it clearer that these are special variables
for combin.lisp.  Plus, `sum` is such a common word that we can't tell
if any occurrence is really the specvar or just a lexical.

In making this change, we do see that some places were using `sum` to
refer to a lexical var instead of the specvar.  This is a big win.

Full testsuite passes.