Rename *combin-var to poly-var in sumsum
We now rename the arg `*combin-var*` for `sumsum` to just `poly-var`.
We don't need the specvar in `sumsum`.

We also forgot to rename `*combin-var*` to `poly-var` in `isum` when
we moved `ipolysum` and friends into `isum`.

Finally, to make this all work, `fbino` needs an extra arg for the
poly-var so that we can rename `*combin-var*` there with `poly-var`.

Full testsuite passes.