Move more functions into sumsum.
We can move `adsum`, `adusum`, `fpolysum`, `fbino`, and `isum` into
`sumsum`, because these are all called by `sumsum` or other functions
called only by `sumsum`.

In particular by doing this and moving `adsum`, and `adusum`, we can
get rid of the special vars `*combin-sum*`, and `*combin-usum*` and
replace them with lexical vars `combin-sum` and  `combin-usum`.

Full testsuite passes.

Not sure I like this huge set of functions all contained in `sumsum`,
but it works and gets rid of two more special vars that don't really
need to be.