Remove $prevfib and $next_lucas
Per the discussion on the mailing list (Jan 28, 2023), these Maxima
variables can be deleted.

For `$prevfib`, just rename it to `prevfib` since all uses are local
to the function `ffib`.  For `$next_lucas`, we can get rid of it
completely because it wasn't really used for anything.

Two more special variables removed from combin.lisp.