Move genexpands as a labels function in expandtimes.
`genexpands` is the only function that needs access to the specvars
`*prods*`, `*negprods*`, `*sums*`, and `*negsums*`.  By moving
`genexpands` into `expandtimes`, these don't need to be special
anymore so we can remove the `declare-top` for these.

However, to make this work, we needed to modify `expandtimes` slightly
by adding a `block` and a `tagbody`.

Full testsuite passes.