Rename former special vars
Since `*prods*`, `*negprods*`, `*sums*`, and `*negsums*` aren't
specvars anymore, rename them to `simp-foo`.  That is we remove the
"*" and add a prefix of "simp-".  The prefix is intended to make them
unique because "prods" and "sums" are such common variable names, so
let's make them unique to prevent any possible current (and future)
confusion with other variables of the same name.

Also did a minor tweak in moving the `block` to just before the
`tagbody` to limit the scope.

Full testsuite passes.